mob next [ci-skip] [ci skip] [skip ci]
lastFile:modules/java/src/test/java/io/github/isitartortrash/approvaltesting/livecoding/OldObjectAssertionsTest.java

+++ modules/java/src/main/kotlin/io/github/isitartortrash/approvaltesting/livecoding/OrderResult.kt
@@ -8,7 +8,7 @@ import java.time.LocalDateTime
 
 data class OrderResult @JsonCreator constructor(
     @JsonProperty("id") val id: String,
-    @JsonProperty("version") val version: Int,
+    @JsonProperty("version") val version: Long,
     @JsonProperty("items") val items: List<ItemResult>,
     @JsonProperty("coupons") val coupons: List<CouponResult> = emptyList(),
     @JsonProperty("orderTimeStamp") val orderTimeStamp: LocalDateTime,
